Moved _digested assets to silver
