+++ hydra/assets/geosampa.py
@@ -1,37 +1,12 @@
-import base64
-from io import BytesIO
 from dagster import (
     AssetExecutionContext,
-    AssetIn,
     AssetsDefinition,
-    MetadataValue,
     Output,
     asset,
 )
-from geopandas import GeoDataFrame
-import matplotlib.pyplot as plt
-import seaborn as sns
 
 from ..resources import GeosampaClient
 from ..config import GeosampaConfig
-
-
-def _get_md_preview_plot(gdf: GeoDataFrame, nome_camada: str) -> str:
-    # Defino o seaborn theme
-    sns.set_theme(rc={'patch.linewidth': 0.1})
-
-    # Ploto o dataframe
-    gdf.plot()
-    plt.axis('off')
-    plt.title(f'Features da camada {nome_camada}')
-
-    # Converto a imagem em um formato decodificável
-    buffer = BytesIO()
-    plt.savefig(buffer, format="png")
-    image_data = base64.b64encode(buffer.getvalue())
-
-    # Converto a imagem em markdown
-    return f"![img](data:image/png;base64,{image_data.decode()})"
 
 
 def __build_raw_asset(name, group_name="geosampa_bronze") -> AssetsDefinition:
@@ -51,7 +26,7 @@ def __build_raw_asset(name, group_name="geosampa_bronze") -> AssetsDefinition:
         context.log.info(f'Camada {name} baixada')
 
         assert camada["type"] == "FeatureCollection"
-        
+
         return Output(
             camada,
             metadata={
@@ -60,55 +35,6 @@ def __build_raw_asset(name, group_name="geosampa_bronze") -> AssetsDefinition:
 
     return _asset
 
-def __build_digested_asset(name, group_name="geosampa_bronze") -> AssetsDefinition:
-    @asset(
-        name=f'{name}_digested',
-        ins={"raw_asset": AssetIn(key=name)},
-        group_name=group_name,
-        io_manager_key="bronze_io_manager",
-        dagster_type=GeoDataFrame,
-    )
-    def _asset(
-        context: AssetExecutionContext,
-        raw_asset: dict
-    ):
-        context.log.info(f'Lendo a camada {name}')
-        gdf = GeoDataFrame.from_features(raw_asset['features'])
-        gdf = gdf.set_crs(raw_asset['crs'].get('properties').get('name'))
-        gdf = gdf.to_crs(epsg=31983)
-
-        context.log.info(f'Camada {name} lida')
-        context.log.info(f'Lendo as configurações da camada {name}')
-        conf = GeosampaConfig.get_asset_config().get('geosampa').get(name)
-        if 'renamed_columns' in conf.keys():
-            context.log.info(f'Renomeando as colunas da camada {name}')
-            gdf = gdf.rename(columns=conf.get('renamed_columns'))
-        else:
-            context.log.info(f'A camada {name} não possui colunas a renomear')
-
-        context.log.info(f'Extraindo metadados da camada {name}')
-
-        # Recebo a imagem de prévia em markdown
-        md_preview = _get_md_preview_plot(gdf, name)
-
-        # Extraio algumas linhas como amostra
-        n = 10 if gdf.shape[0] > 10 else gdf.shape[0]
-
-        peek = gdf.drop(columns=['geometry']).sample(n)
-
-        return Output(
-            gdf,
-            metadata={
-                'núm. linhas': gdf.shape[0],
-                'prévia em gráfico': MetadataValue.md(md_preview),
-                f'amostra de {n} linhas': MetadataValue.md(peek.to_markdown()),
-            })
-
-    return _asset
-
 
 globals().update({asset_: __build_raw_asset(asset_)
-                  for asset_ in GeosampaConfig.get_asset_config().get('geosampa').keys()})
-
-globals().update({f'{asset_}_digested': __build_digested_asset(asset_)
                   for asset_ in GeosampaConfig.get_asset_config().get('geosampa').keys()})
